Add VGA display support
Display mode is autodetected. RISC OS 3.5+ systems (RiscPC and later) default to
VGA. Earlier systems determine monitor type from CMOS.

The current implementation renders to a back buffer in main memory then scales
that to a full size buffer in video memory. This is not the most efficient
solution but appears to run fine on ARM250 hardware.